+++ example/lib/examples/sura_accordian_and_sura_expandable.dart
@@ -43,6 +43,9 @@ class _SuraAccordianAndExpandableState
                 leading: SuraPlatformChecker(androidWidget: CircleAvatar()),
               ),
             ),
+            TextFormField(
+              validator: (value) {},
+            )
           ],
         ),
       ),