+++ lib/src/widgets/sura_loading_dialog.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 
 import 'sura_provider.dart';
 
+///This Widget isn't well tested and shouldn't be use
 class SuraLoadingDialog {
   final bool dismissible;
   final BuildContext context;